Moved the responsibility for rendering NavLight billboard from scenegraph/Model into NavLights.
