+++ src/NavLights.h
@@ -9,6 +9,9 @@
 #include "libs.h"
 #include "Serializer.h"
 #include "json/json.h"
+#include "graphics/RenderState.h"
+#include "graphics/VertexArray.h"
+#include "graphics/VertexBuffer.h"
 
 namespace Graphics { class Renderer; }
 namespace SceneGraph { class Model; class Billboard; }
@@ -39,6 +42,7 @@ public:
 
 	void SetEnabled(bool on) { m_enabled = on; }
 	void Update(float time);
+	void Render(Graphics::Renderer *renderer);
 	void SetColor(unsigned int group, LightColor);
 
 	static void Init(Graphics::Renderer*);
@@ -72,6 +76,10 @@ protected:
 	float m_time;
 	float m_period;
 	bool m_enabled;
+	
+	Graphics::VertexArray m_billboardTris;
+	RefCountedPtr<Graphics::VertexBuffer> m_billboardVB;
+	Graphics::RenderState *m_billboardRS;
 };
 
 #endif
