+++ src/scenegraph/Billboard.cpp
@@ -14,16 +14,16 @@
 
 namespace SceneGraph {
 
-Billboard::Billboard(SceneGraph::Model *model, Graphics::Renderer *r, float size)
+Billboard::Billboard(Graphics::VertexArray& bbVA, Graphics::Renderer *r, float size)
 : Node(r, NODE_TRANSPARENT)
-, m_model(model)
+, m_bbVA(bbVA)
 , m_size(size)
 {
 }
 
 Billboard::Billboard(const Billboard &billboard, NodeCopyCache *cache)
 : Node(billboard, cache)
-, m_model(billboard.m_model)
+, m_bbVA(billboard.m_bbVA)
 , m_size(billboard.m_size)
 {
 }
@@ -42,11 +42,10 @@ void Billboard::Render(const matrix4x4f &trans, const RenderData *rd)
 {
 	PROFILE_SCOPED()
 
-	Graphics::VertexArray& bbVA = m_model->GetBillboardVA();
 	//some hand-tweaked scaling, to make the lights seem larger from distance (final size is in pixels)
 	const float pixrad = Clamp(Graphics::GetScreenHeight() / trans.GetTranslate().Length(), 1.0f, 15.0f);
 	const float size = (m_size * Graphics::GetFovFactor()) * pixrad;
-	bbVA.Add(trans * vector3f(0.0f), vector3f(m_colorUVoffset, size));
+	m_bbVA.Add(trans * vector3f(0.0f), vector3f(m_colorUVoffset, size));
 }
 
 }
